third_party/WebKit/LayoutTests/storage/indexeddb/createIndex-after-failure-expected.txt

Test IndexedDB's basics.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


dbname = "createIndex-after-failure.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore.createIndex('index', 'key', {unique: true})
objectStore.deleteIndex('index')
Expecting exception from objectStore.deleteIndex('index')
PASS Exception was thrown.
PASS code is DOMException.NOT_FOUND_ERR
PASS ename is 'NotFoundError'
Exception message: Failed to execute 'deleteIndex' on 'IDBObjectStore': The specified index was not found.
Now requesting object2
now we wait.
deleteIndexAfterGetError()
Expecting exception from objectStore.deleteIndex('index')
PASS Exception was thrown.
PASS code is 0
PASS ename is 'TransactionInactiveError'
Exception message: Failed to execute 'deleteIndex' on 'IDBObjectStore': The transaction has finished.
Expecting exception from objectStore.deleteIndex('index')
PASS Exception was thrown.
PASS code is 0
PASS ename is 'TransactionInactiveError'
Exception message: Failed to execute 'deleteIndex' on 'IDBObjectStore': The transaction has finished.
PASS successfullyParsed is true

TEST COMPLETE
